Add json rpc call to clear address book
